Refactor form block to use a different structure depending on type, and fix issue with bindings breaking and forms not appearing
