+++ packages/client/src/components/app/blocks/index.js
@@ -1,5 +1,5 @@
 export { default as tableblock } from "./TableBlock.svelte"
 export { default as cardsblock } from "./CardsBlock.svelte"
 export { default as repeaterblock } from "./RepeaterBlock.svelte"
-export { default as formblock } from "./FormBlock.svelte"
+export { default as formblock } from "./form/FormBlock.svelte"
 export { default as chartblock } from "./ChartBlock.svelte"
